requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
# AI & LLM
openai>=1.3.0
langchain>=0.0.340
llama-index==0.9.15
faiss-cpu>=1.7.4
tiktoken>=0.5.1

# Web & Search
googlesearch-python==1.2.3
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.1

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.3

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pytz==2023.3.post1
dateparser==1.1.8
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
